+++ libraries/SysAPI/src/System.cpp
@@ -112,4 +112,9 @@ void System::launch_browser(var::StringView url){
   var::String command = var::String(is_macosx() ? "open" : "start") + " " + url;
   system(command.cstring());
 }
+
+const char * System::get_executable_suffix(){
+  return is_windows() ? ".exe" : "";
+}
+
 #endif